requirements.txt
tensorflow-gpu
keras
numpy
sklearn
matplotlib
mpl_toolkits
lapjv
scipy